Solve the mystery of the mismatched object type
STEP 2. Apply basic OO Principles

i) Start by identifying if each object is doing its job and only its job
to the best of its ability. That is exactly what is happening in Rick's
inventory search tool. The client's specification is given through a
Guitar object but there is no need to use it to compare against the
inventory. Here are some tips for finding the mismatched object type:
- Objects should do what their names indicate
- Each object should represent a single concept
- Unused properties are a dead giveaway

ii) Use ENCAPSULATION for:
- Grouping your application into logical parts
- Avoid duplicate code
- Keep the parts of your code that stay the same from the parts that
change