Add support for 12-strings guitars
1. Add a num_strings property to GuitarSpec class

2. Change the search method in the Inventory class to delegate comparing
the two GuitarSpec objects to the GuitarSpec class, instead of handling
the comparison directly. Please be aware of compare what needs to be
compared. When I was updating the conditional and comparing each guitar
specification to the search_spec; instead of referring to
guitar.guitar_spec I was comparing a Guitar object against a GuitarSpec
object which always will be false

3. Update find_guitar.py to work with the new classes

Use DELEGATION when an object needs to perform a certain task, and
instead of doing the task directly, it asks another object to handle the
task, or sometimes just a part of the task, making it be reused easily.