Emphasize COMMONALITY, before dealing with tests
It keeps the common properties of a Unit outside of the properties Map,
and leaving properties that vary inside the properties Map.

But it violates the DRY principle since there are two ways to access
properties and that is almost certainly going to mean duplicate code
somewhere.

In addition, if a game designer doesn't want to use those, it will
require changes to the Unit class